requirements.txt

GitPython==2.1.11
Pillow==5.2.0
pandas==0.23.4
requests==2.19.1
tqdm==4.26.0
